feat : 내 포인트 조회 기능

+++ src/main/java/avengers/lion/member/domain/Member.java
@@ -41,7 +41,7 @@ public class Member extends BaseEntity {
     private MemberRole role;
 
     @Column(name = "point", nullable = false)
-    private int point;
+    private Long point;
 
     @OneToMany(mappedBy = "member", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<Review> reviews;
@@ -56,7 +56,7 @@ public class Member extends BaseEntity {
         this.password = password;
         this.profileImageUrl = profileImageUrl;
         this.role = role;
-        this.point = 0;
+        this.point = 0L;
     }
 
     public void buyItemByPoint(int price){
